feat(design): promote canonical EmptyState to components/molecules (#12883)
Move the existing canonical EmptyState organism verbatim to
components/molecules per epic #12638 acceptance. Leave a deprecated
re-export shim at the old organisms path so all current consumers,
tests, and stories keep working. Add molecules barrel export.
No visual changes.
